Rename AvoidBoundsErrorsCheck to PreferAtOverSubscriptOperatorCheck
